Don't override package require parameter
Setting `require` to `undef` means it can't be overridden from outside the class to ensure a custom repo is added before trying to install the package. So puppet runs can fail the first time as it can try to install the package before the repo is added.
A little less elegant, but at least it doesn't break puppet environments that use a custom repo.